reduce repetition by adding keychain functions to config struct
